+++ src/just_makeit/templates/c/inc/component_core.h
@@ -42,7 +42,7 @@ extern "C" {
 
 /*<<steps_c_decl>>*/
 
-/*<<getter_setter_decls>>*/
+/*<<getter_setter_decls>>*//*<<serializable_decls>>*/
 
 /*<<property_decls>>*/
 /*<<method_decls>>*//*<<inline_core>>*/